Refactoring: Move get_customer_cards to utils.stripe
